pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knight"
version = "0.1.0"
description = "KNIGHT v1: Brain Neuroimmune Cell State Foundation Model"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Syntropi AI Group"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core ML
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "flash-attn>=2.3.0",
    # Single-cell
    "scanpy>=1.9.6",
    "anndata>=0.10.0",
    "scvi-tools>=1.0.0",
    "muon>=0.1.5",
    # Data
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    # Experiment tracking
    "wandb>=0.16.0",
    # Utilities
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
]
data = [
    "gdown>=4.7.0",
    "requests>=2.31.0",
    "cellxgene-census>=1.8.0",
]

[project.scripts]
knight = "knight.cli:app"

[tool.setuptools.packages.find]
include = ["knight*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
